# perf: add HTTP/2 Server Push and font optimization for Core Web Vitals