--- a/index.html
+++ b/index.html
@@ -40,16 +40,14 @@
     <link rel="dns-prefetch" href="https://fonts.gstatic.com">
     <link rel="dns-prefetch" href="https://egcxzonpmmcirmgqdrla.supabase.co">
     
-    <!-- Preload critical resources for LCP optimization -->
-    <!-- These will be transformed by Vite in production to correct hashed filenames -->
-    <link rel="modulepreload" href="/src/main.tsx" />
-    
-    <!-- Preload icon sprite for faster icon rendering -->
-    <link rel="preload" href="/icons.svg" as="image" type="image/svg+xml" />
-    
     <!-- Preload critical fonts for faster text rendering -->
     <link rel="preload" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" as="style" onload="this.onload=null;this.rel='stylesheet'">
     <noscript><link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap"></noscript>
+    
+    <!-- Preload icon sprite if using SVG icons -->
+    <link rel="preload" href="/icons.svg" as="image" type="image/svg+xml" fetchpriority="low">
+    
+    <!-- Vite will inject modulepreload links for critical chunks automatically -->
     
     <!-- Prefetch likely next navigations -->
     <link rel="prefetch" href="/api/contributors" />